Create design-tokens-gitness
For the design team, we need a dedicated folder in GitHub for testing purposes. We will create design tokens using the Token Studio plugin for Figma and then push them to Git. This environment will serve as a place to store and test these tokens until we figure out the best way to deliver all new tokens to the code properly.